fix google-runtime-int

+++ src/odb/src/def/cdef/xdefrReader.cpp
@@ -1176,7 +1176,7 @@ int defrLineNumber()
   return DefParser::defrLineNumber();
 }
 
-long long defrLongLineNumber()
+int64_t defrLongLineNumber()
 {
   return DefParser::defrLongLineNumber();
 }
@@ -1256,7 +1256,7 @@ void defrDisablePropStrProcess()
   DefParser::defrDisablePropStrProcess();
 }
 
-void defrSetNLines(long long n)
+void defrSetNLines(int64_t n)
 {
   DefParser::defrSetNLines(n);
 }
